Refactor the blog component
