+++ src/component/Shared/Features/FeaturedPost.js
