+++ src/component/Shared/Features/MainFeaturedPost.js
