+++ src/component/Shared/Main/Main.js
@@ -4,7 +4,7 @@ import { makeStyles } from "@material-ui/core/styles";
 import Grid from "@material-ui/core/Grid";
 import Typography from "@material-ui/core/Typography";
 import Divider from "@material-ui/core/Divider";
-import Markdown from "markdown-to-jsx";
+import Markdown from "./Markdown";
 
 const useStyles = makeStyles((theme) => ({
   markdown: {
@@ -23,7 +23,9 @@ function Main(props) {
         {title}
       </Typography>
       <Divider />
-      <Markdown>{posts}</Markdown>
+      {posts.map((post) => {
+        return <Markdown>{post}</Markdown>;
+      })}
     </Grid>
   );
 }
